Use in64_t instead of uint32_t for Cesium ion asset IDs.

+++ Cesium3DTilesSelection/src/IonRasterOverlay.cpp
@@ -22,7 +22,7 @@ namespace Cesium3DTilesSelection {
 
 IonRasterOverlay::IonRasterOverlay(
     const std::string& name,
-    uint32_t ionAssetID,
+    int64_t ionAssetID,
     const std::string& ionAccessToken,
     const RasterOverlayOptions& overlayOptions)
     : RasterOverlay(name, overlayOptions),
